refactor: move ScopedCommand into src
